refactor(frontend): Rename Ethereum fee store context
